It builds!

+++ trikotFoundation/src/tvosMain/kotlin/com/mirego/trikot/foundation/concurrent/dispatchQueue/iOSGlobalDispatchQueue.kt
@@ -1,9 +1,9 @@
 package com.mirego.trikot.foundation.concurrent.dispatchQueue
 
 import com.mirego.trikot.foundation.concurrent.freeze
+import platform.darwin.DISPATCH_QUEUE_PRIORITY_HIGH
 import platform.darwin.dispatch_async
 import platform.darwin.dispatch_get_global_queue
-import platform.darwin.DISPATCH_QUEUE_PRIORITY_HIGH
 
 open class iOSGlobalDispatchQueue : DispatchQueue {
     override fun isSerial() = false
@@ -11,7 +11,7 @@ open class iOSGlobalDispatchQueue : DispatchQueue {
     override fun dispatch(block: DispatchBlock) {
         freeze(block)
 
-        dispatch_async(dispatch_get_global_queue(DISPATCH_QUEUE_PRIORITY_HIGH, 0), freeze {
+        dispatch_async(dispatch_get_global_queue(DISPATCH_QUEUE_PRIORITY_HIGH.toLong(), 0UL), freeze {
             runQueueTask(block)
         })
     }
